third_party/blink/web_tests/http/tests/images/feature-policy-unoptimized-images-cached-image-expected.txt

CONSOLE ERROR: Feature policy violation: unoptimized-lossy-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: unoptimized-lossy-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: unoptimized-lossy-images is not allowed in this document.
